Create Room:
Added the functionality to add a room to an users "rooms" collection.

-Added a new Fragment in the Mainactivity that creates a Room with
following Attributes:
Name, owner email and Category

Category are types of rooms that can be choosen for example, Office,
Sleepingroom, kitchen ... .

- The Mainactivity contains a Recyclerview that displays all the rooms
that a user currently has. There is special layout to display these.
Later on is planed that the user can add devices to the rooms.
Further he can control these devices and get information about them.

- Did some bugfixes

TODO:
- Devices
- Cloudfunctions