Create and manage the Devices:
Added the functionality to add a device and manage this with a dashboard.

- The new Dashboard Activity lists all devices that are in a room. Each
device that is listed has shown its name it current Power-status and
Device-type. all the information that is displayed updates automatically
if there are any changes on the database.

- The Dashboard Activity includes a Fragment that adds a new device to
the current room that is selected. The user selects a name and the
category of the device

- Each device can be managed in the "Manage device" fragment that can be
open by clicking on the blue dashboard icon. the user will late be abel
to start the Washing-machine or the toaster, but this will be done with
cloud-functions.

- Fixed some Bugs

TODO:
- Add Cloud-functions to the Manage device fragment that will be triggered with the buttons. For example that the temperature of the
washing-machine will be changed every 5 seconds and this will be displayed. For the toaster a timer will be set that activates the toaster and after 30 seconds it will turn itself off.